refactor(16216): restructure components
