+++ hivemq-edge/src/frontend/src/modules/EdgeVisualisation/components/nodes/NodeBridge.tsx
@@ -6,12 +6,12 @@ import { Bridge } from '@/api/__generated__'
 import { ConnectionStatusBadge } from '@/components/ConnectionStatusBadge'
 import logo from '@/assets/hivemq/05-icon-hivemq-bridge-extension.svg'
 
-import GenericNode from './GenericNode.tsx'
+import NodeWrapper from '../parts/NodeWrapper.tsx'
 
 const NodeBridge: FC<NodeProps<Bridge>> = ({ data: bridge }) => {
   return (
     <>
-      <GenericNode p={3}>
+      <NodeWrapper p={3}>
         <VStack>
           <HStack w={'100%'}>
             <Image boxSize="20px" objectFit="scale-down" src={logo} />
@@ -21,7 +21,7 @@ const NodeBridge: FC<NodeProps<Bridge>> = ({ data: bridge }) => {
             <ConnectionStatusBadge status={bridge.bridgeRuntimeInformation?.connectionStatus?.status} />
           </Box>
         </VStack>
-      </GenericNode>
+      </NodeWrapper>
       <Handle type="source" position={Position.Top} id="Top" isConnectable={true} />
     </>
   )
